project/ctfs/requirements.txt
gitPython
shutil